Add shardingsphere-sql-translator-provider module (#17648)
* Move translator into api module

* Add shardingsphere-sql-translator-provider module

* Add shardingsphere-sql-translator-provider module
