+++ shardingsphere-kernel/shardingsphere-sql-translator/shardingsphere-sql-translator-provider/shardingsphere-sql-translator-native-provider/src/main/resources/META-INF/services/org.apache.shardingsphere.sqltranslator.spi.SQLTranslator
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.sqltranslator.spi.type.NativeSQLTranslator
+org.apache.shardingsphere.sqltranslator.natived.NativeSQLTranslator
